moves livewire files
