+++ app/Livewire/SearchBar.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Filament\Components;
+namespace App\Livewire;
 
 use Livewire\Component;
 
@@ -43,6 +43,6 @@ class SearchBar extends Component
 
     public function render()
     {
-        return view('components.search-bar');
+        return view('livewire.search-bar');
     }
 }
