+++ config/livewire.php
@@ -13,7 +13,7 @@ return [
     |
     */
 
-    'class_namespace' => 'App\\Filament\\Components',
+    'class_namespace' => 'App\\Livewire',
 
     /*
     |---------------------------------------------------------------------------
@@ -26,7 +26,7 @@ return [
     |
     */
 
-    'view_path' => resource_path('views/components'),
+    'view_path' => resource_path('views/livewire'),
 
     /*
     |---------------------------------------------------------------------------
