+++ resources/views/all-troves-table.blade.php
